Enum route
The enum route was the simplest way to get the functionality working. ChessPiece is an enum type where each value holds one of the concrete types for the Piece trait. The board handles ChessPiece enums in the grid to forgo the dynamic trait type issues. This route may need to be revisited at the end depending of performance.