MVC implementation
Decided to follow the model, view, controller route and store the controller and view inside of the game struct for now. This makes more logical sense for the flow of the game where the controller handles the communication with the player for input and the view is responsible for displaying the updated board data. Removed the need for player structs.