Started on the piece movement validation
The tuple containing 2 positions has been changed to a struct containing the current and target positions, to make it more readable. Made a change to the get_current_board function in the board struct to return a copy of the board rather than a reference. Started on the logic in 'game' which performs the validation on the move input. The piece structs, view and controller have been updated to reflect any changes to the game.