Finished the remaining piece move logic
Implemented the move data logic for the pawn, king and knight. Renamed the move data util module to piece util. Reverted the changed to 'attacks' in the move data struct but after initial tests I may have the make more changes to the struct to work for the pawn movements.