Updated move data with a separation of attacking and valid move positions
Attacks and valid move positions are now separate vectors in the move data struct. This is to help solve the problem with pawns having separate positions for both. This will need looking at in future as it is duplicated data. Game struct has been updated to hold a state object containing the constantly changing values.